RIDER-132995 refactoring: move deployment tree to a separate class
